hikey960: Remove wl1271-nvs.bin firmware file from hikey960 build
As reported by Guodong, when using the wl1271-nvs.bin file with
hikey960, we end up with all devices using the same MAC address
dead:beef:0000

Guodong suggested that we avoid including this config file
with hikey960 as a workaround until he can generate a specific
nvs file for WL1837MOD.

This removes the file completely from the build.

I've validated that with the file removed, wifi still functions.

Change-Id: Iebb99cfb1e9ff0029a5a48699bbe0cbd7190c3ba
Signed-off-by: John Stultz <john.stultz@linaro.org>